requirements.txt
alembic==1.13.1
emails==0.6
fastapi==0.109.2
fastapi-pagination==0.12.24
jinja2==3.1.3
loguru==0.7.2
msgspec==0.18.6
passlib[bcrypt]==1.7.4
phonenumbers==8.13.37
pydantic-extra-types==2.7.0
pydantic-settings==2.2.0
pydantic[email]==2.6.1
PyJWT==2.8.0
python-multipart==0.0.9
SQLAlchemy==2.0.27
uvicorn==0.27.1
